Restore command_user_describe_onprem.go as its own file
The on-prem describe belongs in its per-verb file, matching the repo's
one-file-per-verb convention (and the pre-migration layout, restored
byte-identical). command_user_onprem.go keeps only the on-prem command
constructor and its output struct.
